fix(*): fix critical issue introduced in previously published v1.5.0
In a since-unpublished version of atom-ts-transpiler which was pushed
out on 26 Mar 2018 at around 19:00 UTC, a critical issue was introduced
which prevented files from being transpiled correctly if they were not
located on the package root directory (i.e. if they were in a src or lib
subdirectory.) This patch addresses this issue and makes v1.5.0 ready
for republication.

Closes #36.